feat!: Use separate methods for different correlation coefficients
