+++ .github/workflows/TestSuite.yml
@@ -7,12 +7,12 @@ on:
     branches: [ main ]
 
 env:
-  DEVELOPER_DIR: /Applications/Xcode_14.0.app
+  DEVELOPER_DIR: /Applications/Xcode_16.1.app
 
 jobs:
   macOS:
-    runs-on: macos-12
-    
+    runs-on: macos-15
+
     steps:
     - uses: actions/checkout@v3
     - name: Run macOS Tests
@@ -25,6 +25,6 @@ jobs:
     - uses: actions/checkout@v3
     - uses: swift-actions/setup-swift@v1
       with:
-        swift-version: '5.7'
+        swift-version: '5.9'
     - name: Run Linux Tests
       run: swift test -c release --parallel
