--- Local Test Helpers/docker-qa.yml
@@ -1,17 +1,0 @@
-version: "3.8"
-services:
-  linux_5_7:
-    container_name: linux_5_7
-    image: swift:5.7.0-amazonlinux2
-    volumes:
-      - ./:/statkit
-    working_dir: /statkit
-    command: swift test -c release --parallel
-  
-  linter:
-    container_name: linter
-    image: norionomura/swiftlint:0.46.5_swift-5.5.1
-    volumes:
-      - ./:/statkit
-    working_dir: /statkit
-    command: swiftlint lint --strict
